feat: introduce abort method for immediate pathfinding termination and result return
Refactored the Pathfinder classes to include a new `abort()` method, enabling immediate termination of the pathfinding process and returning the current result. Additionally, the specific Pathfinder implementation has been refactored to operate in a tick-wise manner.
