+++ pathetic-api/src/main/java/org/patheloper/api/pathing/result/PathState.java
@@ -2,6 +2,8 @@ package org.patheloper.api.pathing.result;
 
 public enum PathState {
 
+  /** The pathfinding process was aborted */
+  ABORTED,
   /** Pathing failed to start, typically due to an invalid start or end position. */
   INITIALLY_FAILED,
   /** The Path was successfully found */
