+++ pathetic-model/src/main/java/org/patheloper/model/pathing/pathfinder/Depth.java
@@ -1,0 +1,17 @@
+package org.patheloper.model.pathing.pathfinder;
+
+import lombok.Getter;
+
+@Getter
+public class Depth {
+
+  private int depth;
+
+  public Depth(int depth) {
+    this.depth = depth;
+  }
+
+  public void increment() {
+    depth++;
+  }
+}
